Logger Facade
This change adds a Logger abstraction that allows users to provide
a logger facade allowing for control over how request and response
objects are logged during processing.  The default Logger is based on
SLF4J.  I am aware that core should try to remain dependency free, but
I think it's ok to add a logger here, since it's something most, if
not all users, are going to do.

Additonal Changes include:

* Refactored Headers on Request to List from Array to be consistent with the Response
* Refactored HttpResponse to cache the response once `toByteArray` is called, so
`toByteArray` can be called as many times as necessary while the response is open.
This was required to support logging of the response body
* Adding debug log statements throughout AbstractMethodHandler and BlockingMethodHandler
* Adding debug log statements throughout FeignContract and AnnotationContract
* Additional Unit Tests to ensure the Response can be read more than once.
* Unit and Integration Tests