Adding request retry support
Fixes #38

This change includes a `Retry` interface, which is responsible for
determining if the provided callback method should be retried.

I've included two implementations:

* NoRetry - does not retry at all, returns the result of the callback
* ConditionalRetry - relies on a `RetryCondition` to determine if the
  callback should be retried.

`RetryCondition` is a custom `Predicate`-like object that can be
logically 'OR', 'AND', and 'NOT' combined with other conditions to
create more complex conditions.  I've included a few simple
condition implementations based on our current history:

* StatusCodeCondition - inspect the result and retry based on the
  status code
* ExceptionCondition - retry if a given exception is thrown

`RetryCondition` is a functional interface, so conditions can be
created via lambda functions as well.  This should accomodate
almost any condition a user can think of.

Lastly, when using the `ConditionalRetry` and an exception occurs,
the `ExceptionHandler` will be invoked before passing the result
to the Condition.  This allows users to convert the internal
FeignExceptions into business exceptions and configure their retry
logic using their own exceptions, in addition to our Feign based
exceptions.

This approach will provide more flexibility to our users and has
the added benefit of simplifying a lot of the workarounds being
maintained in core feign.