GH-48: User Supplied Template Expansion
Fixes: #48

This change refactored the `UriTemplate` logic to delegate the expansion
logic for each expression to a new `ExpressionExpander` interface.
This interface can then be specified on a `@Param` annotation,
overriding our internal expander resolution.

To support this, much of the internal workings of the `UriTemplate`
has been refactored.  In addition, `FeignContract` has been updated
to determine the appropriate expander at Contract time.

Additional Changes include:

* TCK Passes
* Formatting and Javadoc
* Test Coverage