Gh 47 object expansion (#51)
* GH-47: Object Expansion

Fixes #47

Adds support for expanding simple objects as part of a Uri Template.
This feature uses the Java Beans api to inspect the object, relying
on the read methods for the bean's properties to determine what
information should be included in the expansion.

Nested object are also expanded, their names will be prefixed
with a dot ".".  Example "person.address".

Nested Lists and Maps are delegated to their respective expanders.
All other values are expanded using the `SimpleExpander`.

This implementation will fit for the most common simple use cases.
Complex use cases such as generic variables, interfaces, and
inheritance have not been tested and considered unsupportted at
this time.

Lastly, Dates are expanded using thier `toString` representation.
We may want to consider a date formatting option in the future.