Update SimpleLogger and MethodHandler Loggers to be assigned to Target (#61)
This change updates the SimpleLogger's `log` instance to be assigned
to a user supplied name, allowing users to control what object the
log messages are attributed to.

In addition, `AbstractTargetMethodHandler`'s internal Logger has been
updated to be targeted to the `Target` instead of itself.  This allows
all internal Feign logging during method handling to be attributed to
the Target instance instead of Feign.